User Navigation:
Made a user button and components which displays the user avatar.
Also created a page for this, but has no design and content yet.
Also finised designing pick a minifig page with changing the layout
of the page (less text and images closer together) and a button to
submit the lego minifigure.