Firt implementation of dynamic data:
Created the lego-api project with an index.js file where for now the data
will be stored. The data will be transmitted with a get request.
The lego-shop project receives this data with a fetch api, on all of the
home components and the connected views.

The data on the home components are randomly changing by every refresh, and
are coming from the total of items in it's const.

Made every home component await showing the data untill it has received it,
by a loader.

Changed the idea of pick a minifig to just pick a minifig head, but havn't
implemented it yet in the code. The person can select a head and give it a
name, and set it as an avatar. Also, to fill the yourcollection page, the
person can make a item a favorite and this will get saved into the database
and into the collection section.

This will be the last commit before creating an online database, and before the
reset of windows and by that my whole laptop.

Pushed data will make the app not work untill the database is created and connected.