Implementation of the use of localstorage for the collection:
First we put the add and remove from favorites heart to its own
component and we use it in the legoItem and the legoComingSoon.
The comments explain how it works.
For next time, some naming conventions are different for the
legoItem and legoComingSoon, so change it so all of the data
gets displayed in the collection.
yaaay